+++ src/frontend/apps/standalone_site/src/features/Contents/index.tsx
@@ -5,4 +5,5 @@ export { default as ContentsRouter } from './components/ContentsRouter/ContentsR
 export { default as ContentsShuffle } from './components/ContentsShuffle/ContentsShuffle';
 export { default as ManageAPIState } from './components/ManageAPIState/ManageAPIState';
 export { ClassRoomRouter, ClassRooms } from './features/ClassRoom/';
+export { LiveRouter, Lives } from './features/Live/';
 export { VideoRouter, Videos } from './features/Video/';